// Frontend/SolarNet/src/app/components/page-ident/main-ident/advertisment-content-ident/models/Advertisment.ts
export class Adv{
        advertismentTitle: string = "";
        categoryId: number = 0;
        date: string = "";
        description: string = "";
        id: number = 0;
        images: string[] = [];
        price: number = 0;
        userName: string = "";
        city: string = "";
        commentsCount: number = 0;
        userId: string = "";
        userPhone: string = "";
        deleteReason: string = "";
        stetus: string = "";
}